requirements.txt
chumpy==0.70
einops==0.8.0
fire==0.5.0
kornia==0.6.12
libigl==2.4.1
lmdb==1.4.1
loguru==0.7.0
manopth==0.0.1
matplotlib==3.6.2
numpy==1.22.4
opencv_python==4.6.0.66
opencv_python_headless==4.7.0.72
pandas==1.5.1
pycocotools==2.0.6
pyglet==2.0.0
pyrender==0.1.45
PyYAML==6.0
PyYAML==6.0.2
scikit_image==0.16.2
scikit_learn==1.1.3
scipy==1.9.1
setuptools==65.5.0
tabulate==0.9.0
torch==2.4.1+cu118
torchvision==0.19.1+cu118
tqdm==4.65.0
tqdm==4.64.1
trimesh==3.16.0
yacs==0.1.8
~cikit_image==0.16.2
